Added OptimizelyDecisionTest
