+++ OptimizelySDK.Tests/OptimizelyDecisions/OptimizelyDecisionTest.cs
@@ -1,4 +1,20 @@
-﻿using Moq;
+﻿/* 
+ * Copyright 2020, Optimizely
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+using Moq;
 using NUnit.Framework;
 using OptimizelySDK.ErrorHandler;
 using OptimizelySDK.Logger;
@@ -63,6 +79,18 @@ namespace OptimizelySDK.Tests.OptimizelyDecisions
             Assert.AreEqual(optimizelyDecision.Reasons, new List<string>());
             Assert.AreEqual(optimizelyDecision.RuleKey, "experiment");
             Assert.True(optimizelyDecision.Enabled);
+        }
+
+        [Test]
+        public void TestNewDecisionReasonWithDecideAllOptions()
+        {
+            var decisionReasons = DefaultDecisionReasons.NewInstance(new List<OptimizelyDecideOption>() { OptimizelyDecideOption.INCLUDE_REASONS });
+            decisionReasons.AddError(DecisionMessage.Reason(DecisionMessage.FLAG_KEY_INVALID, "invalid_key"));
+            Assert.AreEqual(decisionReasons.ToReport()[0], "No flag was found for key \"invalid_key\".");
+            decisionReasons.AddError(DecisionMessage.Reason(DecisionMessage.VARIABLE_VALUE_INVALID, "invalid_key"));
+            Assert.AreEqual(decisionReasons.ToReport()[1], "Variable value for key \"invalid_key\" is invalid or wrong type.");
+            decisionReasons.AddInfo("Some info message.");
+            Assert.AreEqual(decisionReasons.ToReport()[2], "Some info message.");
         }
 
     }
